Resolve stack value issue, get callback reference working. Async example works!
This took a lot of reading up and learning by example to do, but I finally have async handling working (for the most part).

There were two major roadblocks to getting this to work:
1. Right now I'm limited to callbacks done in the async completion context, after all the async work is done.
2. I wasn't allocating my carrier values to the heap memory, so that context was lost on calling the async work.
3. Similar to 2, my napi callback function required me to create a reference to it, and passing the reference to the async carrier to retrieve the callback value later.

With this breakthrough, I can at least set up a nondeterministic progress bar, but won't be able to relay progress updates to the renderer during the patching progress yet. That will require a little more work on my end to support, but this is still huge for me nonetheless.