Support salts for randomizer hashing
If a player has the necessary build tools installed, they can simply
build against a version of the randomizer and debug the application in
order to cheat. To counteract this, each release of the randomizer,
starting with v1.0.0, will use a salt that will not be checked into the
repo. This is admittedly security by obscurity, but should weed
out anyone immediately looking to cheat without putting in
effort.