make canEscape() more accurate
The item to place should be considered when checking if you can leave a room.

For instance, if Wave Beam is being placed in Chapel of the Elders, it isn't being considered when checking if you can leave
the room, so it then becomes impossible for it to be placed in its own vanilla location.