Make item dependency checking more accurate
In addition to checking if the location can be filled, a second search
is now performed to verify that the player can safely leave the
encompassing region from its original entry point in the search. This
is to handle cases where the player can enter the region, but needs
additional items to escape, such as Ventilation Shaft.